requirements.txt
google-adk>=1.0.0
litellm>=1.0.0
python-dotenv>=1.0.0
uvicorn>=0.20.0
aiosqlite>=0.19.0
